feat: update climatetrace importer to import all point data
